LFS-752: Create script to package nightly file export Move export to seperate package and set to nightly
